pyproject.toml
[project]
name = "dteam-django-practical-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Volodymyr Zherdetskyi",email = "91687684+vzhV@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "xhtml2pdf (>=0.2.17,<0.3.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"

